dev(rest/client) created clientHeader const key in rest pkg, also modified some of the tests related to the rest pkg in the project
