+++ cmd/inspr/cli/delete_test.go
@@ -356,7 +356,7 @@ func Test_deleteApps(t *testing.T) {
 	handler := func(w http.ResponseWriter, r *http.Request) {
 		data := models.AppQueryDI{}
 		decoder := json.NewDecoder(r.Body)
-		scope := r.Header.Get("Scope")
+		scope := r.Header.Get(rest.HeaderScopeKey)
 
 		err := decoder.Decode(&data)
 		if err != nil {
@@ -436,6 +436,7 @@ func Test_deleteChannels(t *testing.T) {
 
 	handler := func(w http.ResponseWriter, r *http.Request) {
 		data := models.ChannelQueryDI{}
+		scope := r.Header.Get(rest.HeaderScopeKey)
 		decoder := json.NewDecoder(r.Body)
 
 		err := decoder.Decode(&data)
@@ -443,8 +444,7 @@ func Test_deleteChannels(t *testing.T) {
 			fmt.Println(err)
 		}
 
-		// TODO: revise, removed `data.Scope != "appParent" ||`
-		if data.ChName != "ch1" {
+		if scope != "appParent" || data.ChName != "ch1" {
 			rest.ERROR(w, ierrors.NewError().Message("error test").Build())
 			return
 		}
@@ -518,7 +518,7 @@ func Test_deleteCTypes(t *testing.T) {
 	handler := func(w http.ResponseWriter, r *http.Request) {
 		data := models.ChannelTypeQueryDI{}
 		decoder := json.NewDecoder(r.Body)
-		scope := r.Header.Get("Scope")
+		scope := r.Header.Get(rest.HeaderScopeKey)
 
 		err := decoder.Decode(&data)
 		if err != nil {
@@ -599,13 +599,14 @@ func Test_deleteAlias(t *testing.T) {
 	handler := func(w http.ResponseWriter, r *http.Request) {
 		data := models.AliasQueryDI{}
 		decoder := json.NewDecoder(r.Body)
-		scope := r.Header.Get("Scope")
+		scope := r.Header.Get(rest.HeaderScopeKey)
 
 		err := decoder.Decode(&data)
 		if err != nil {
 			fmt.Println(err)
 		}
 
+		fmt.Println(scope)
 		if scope != "appParent" || data.Key != "alias_name" {
 			rest.ERROR(w, ierrors.NewError().Message("error test").Build())
 			return
