+++ pkg/client/client.go
@@ -34,8 +34,7 @@ func NewAppClient() *Client {
 			BaseURL("http://unix").
 			HTTPClient(transports.NewUnixSocketClient(envAddr)).
 			Encoder(json.Marshal).
-			Decoder(request.JSONDecoderGenerator).
-			Build(),
+			Decoder(request.JSONDecoderGenerator),
 	}
 }
