+++ pkg/rest/constants.go
@@ -1,0 +1,6 @@
+package rest
+
+const (
+	// HeaderScopeKey is used to obtain or to change the scope in the http request header, usually used in the handlers of the api package and the client of the controller package.
+	HeaderScopeKey = "Scope"
+)
