+++ pkg/rest/request/request_test.go
@@ -366,67 +366,3 @@ func TestClient_handleResponseErr(t *testing.T) {
 		})
 	}
 }
-
-func TestClient_routeToURL(t *testing.T) {
-	type args struct {
-		route string
-	}
-	tests := []struct {
-		name string
-		c    *Client
-		args args
-		want string
-	}{
-		{
-			name: "basic testing",
-			c: &Client{
-				baseURL: "http://test",
-			},
-			args: args{
-				route: "/route",
-			},
-			want: "http://test/route",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := tt.c.routeToURL(tt.args.route); got != tt.want {
-				t.Errorf("Client.routeToURL() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestJSONDecoderGenerator(t *testing.T) {
-	type args struct {
-		value interface{}
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{
-			name: "decoder creation",
-			args: args{
-				value: "hello",
-			},
-			want: "hello",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			encoded, _ := json.Marshal(tt.args.value)
-			gotDecoder := JSONDecoderGenerator(bytes.NewBuffer(encoded))
-			var got string
-			err := gotDecoder.Decode(&got)
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("JSONDecoderGenerator() = %v, want %v", got, tt.want)
-			}
-			if err != nil {
-				t.Error("error in decoding")
-			}
-		})
-	}
-}
